pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reposec"
version = "0.1.0"
description = "Reusable security audit tool for any repository"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "DevOpsCelstn" }]
keywords = ["security", "audit", "cli", "static-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pathspec>=0.11",
]

[project.scripts]
reposec = "reposec.cli:app"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]

[project.entry-points."reposec.rules"]
# Third-party rule packages register here

[tool.hatch.build.targets.wheel]
packages = ["src/reposec"]

[tool.pytest.ini_options]
testpaths = ["tests"]
